BountySchema + BountyListSchema, getBountyWithDetails
