+++ apps/web/app/app.dub.co/(dashboard)/[slug]/(ee)/program/bounties/bounty-list.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import useWorkspace from "@/lib/swr/use-workspace";
-import { BountyProps } from "@/lib/types";
+import { BountyListProps } from "@/lib/types";
 import { AnimatedEmptyState } from "@/ui/shared/animated-empty-state";
 import { Button } from "@dub/ui";
 import { fetcher } from "@dub/utils";
@@ -18,7 +18,7 @@ export function BountyList() {
     data: bounties,
     isLoading,
     error,
-  } = useSWR<BountyProps[]>(
+  } = useSWR<BountyListProps[]>(
     workspaceId ? `/api/bounties?workspaceId=${workspaceId}` : null,
     fetcher,
     {
