+++ apps/web/lib/types.ts
@@ -25,6 +25,7 @@ import {
 } from "./folder/constants";
 import { WEBHOOK_TRIGGER_DESCRIPTIONS } from "./webhook/constants";
 import {
+  BountyListSchema,
   BountySchema,
   BountySchemaExtended,
   BountySubmissionExtendedSchema,
@@ -526,7 +527,7 @@ export type GroupProps = z.infer<typeof GroupSchema>;
 export type GroupExtendedProps = z.infer<typeof GroupSchemaExtended>;
 
 export type BountyProps = z.infer<typeof BountySchema>;
-
+export type BountyListProps = z.infer<typeof BountyListSchema>;
 export type BountyExtendedProps = z.infer<typeof BountySchemaExtended>;
 
 export type PartnerBountyProps = z.infer<typeof PartnerBountySchema>;
