+++ apps/web/lib/webhook/sample-events/bounty-created.json
@@ -6,10 +6,5 @@
   "startsAt": "2025-08-01T17:34:00.000Z",
   "endsAt": "2025-09-01T17:34:00.000Z",
   "rewardAmount": 10000,
-  "submissionRequirements": null,
-  "performanceCondition": {
-    "attribute": "totalConversions",
-    "operator": "gte",
-    "value": 20
-  }
+  "submissionRequirements": null
 }
